🐛 Fix: refactor profile page
