+++ src/components/Profile/ProfileCard.tsx
@@ -1,0 +1,5 @@
+const ProfileCard = ({ children }) => {
+	return <div className="bg-gray-50 rounded-lg p-5">{children}</div>;
+};
+
+export default ProfileCard;
